pyproject.toml
[tool.poetry]
name = "toronto-man-translator"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.13.0"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
